Chop nameprefix if too long
Nameprefix cannot be longer than 32 characters and given situation with
long(ish) component name it's easy to cross 32 characters given we also
have suffix `service-role` (this came out of testing via Hypotesis in
another Terraform module which relies on this module).

Given this is name prefix for a specific resource (either role or role
policy) I removed the suffix and added condition in which, if passed
service name will cross 32 characters it'll get chopped at 32
characters.